Refined the 6th step
So the 'algorithm' that 'decides' what icon and what alternative text has to be used depending on the forecast, has moved to a function in a separate .js file. Two of the components that were using the algorithm are now calling the function so this keeps the total code shorter and better organised. In the process, I corrected a couple of faults on the code.